Broke the index into sections to make it easier to extract components and seperate logic from input.
